+++ app/about/wakatime.tsx
@@ -1,34 +1,10 @@
-import { z } from "zod";
+"use client";
 
-import { env } from "~/env.mjs";
+import { useFetch } from "~/hooks/use-fetch";
+import { WakaTimeSchema } from "~/lib/schema";
 
-const WakaTimeResponseSchema = z.object({
-  data: z.object({
-    languages: z.array(
-      z.object({
-        name: z.string(),
-        text: z.string(),
-        percent: z.number(),
-      }),
-    ),
-  }),
-});
-
-// TODO: make this work, dont render on build
-export default async function WakaTime() {
-  const rawData = await fetch(
-    "https://wakatime.com/api/v1/users/current/stats/last_7_days",
-    {
-      headers: {
-        Authorization: `Basic ${Buffer.from(
-          env.WAKATIME_SECRET_API_KEY || "",
-        ).toString("base64")}`,
-      },
-      next: { revalidate: 86400 },
-    },
-  ).then((res) => res.json());
-  const { data } = WakaTimeResponseSchema.parse(rawData);
-
+// TODO: make it better
+export default function WakaTime() {
   return (
     <div className="mt-6 flex w-full flex-col rounded-xl border bg-background p-4">
       <p className="inline-flex items-center gap-1 font-medium">
@@ -37,13 +13,25 @@ export default async function WakaTime() {
           (Last 7 days)
         </span>
       </p>
-      <div className="mt-3 flex flex-col">
-        {data.languages.slice(0, 5).map(({ name, text, percent }, i) => (
-          <span key={i}>
-            {name}: {text} {percent}%
-          </span>
-        ))}
-      </div>
+      <Data />
+    </div>
+  );
+}
+
+function Data() {
+  const { data } = useFetch("/api/wakatime", WakaTimeSchema);
+
+  if (!data) {
+    return <p>Loading...</p>;
+  }
+
+  return (
+    <div className="mt-3 flex flex-col">
+      {data.languages.slice(0, 5).map(({ name, text, percent }, i) => (
+        <span key={i}>
+          {name}: {text} {percent}%
+        </span>
+      ))}
     </div>
   );
 }
